Code clean up and added texture scaling

+++ ESPStream.h
@@ -36,7 +36,7 @@ typedef struct {
 int build_fd_sets(int socket, fd_set *read_fds, fd_set *write_fds);
 int connect_serv (server_con *sc);
 server_con espstream_init(char *ip_address, uint8_t port);
-uint32_t *espstream_get_image(server_con *sc, int *imgWidth, int *imgHeight, int *size);
+uint8_t *espstream_get_image(server_con *sc, int *imgWidth, int *imgHeight, int *size);
 void espstream_cleanup();
 
 #endif
